+++ dto/authorization_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019 Authlete, Inc.
+// Copyright (C) 2019-2020 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -74,6 +74,12 @@ type AuthorizationResponse struct {
 
 	//
 	UserInfoClaims string `json:"userInfoClaims,omitempty"`
+
+	//
+	Resources []string `json:"resources,omitempty"`
+
+	//
+	Purpose string `json:"purpose,omitempty"`
 
 	//
 	ResponseContent string `json:"responseContent,omitempty"`