+++ CHANGES.md
@@ -7,6 +7,10 @@ CHANGES
 - `AuthorizationIssueRequest` struct
     * Added `IdtHeaderParams` field.
 
+- `AuthorizationResponse` struct
+    * Added `Resources` field.
+    * Added `Purpose` field.
+
 v1.0.5 (2019-09-25)
 -------------------